requirements.txt
alembic==1.8.1
anyio==3.6.1
apache-airflow==2.3.4
apache-airflow-providers-common-sql==1.1.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-microsoft-mssql==3.2.0
apache-airflow-providers-sqlite==3.2.0
apispec==3.3.2
argcomplete==2.0.0
attrs==22.1.0
Authlib==0.15.5
azure-common==1.1.28
azure-core==1.25.1
azure-identity==1.9.0
azure-mgmt-core==1.3.2
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==21.0.0
azure-mgmt-subscription==3.0.0
Babel==2.10.3
bcrypt==3.2.2
bidict==0.22.0
black==22.8.0
blinker==1.5
boto3==1.20.54
botocore==1.23.54
Brotli==1.0.9
cachelib==0.9.0
cachetools==5.2.0
cattrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
clickclick==20.10.2
colorama==0.4.5
colorlog==4.8.0
commonmark==0.9.1
connexion==2.14.0
cron-descriptor==1.2.31
croniter==1.3.5
cryptography==36.0.2
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.19
email-validator==1.2.1
eventlet==0.33.0
exceptiongroup==1.0.0rc8
Flask==2.2.2
Flask-AppBuilder==4.1.3
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-Compress==1.12
Flask-Gravatar==0.5.0
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.5
Flask-Session==0.4.0
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
google-api-core==2.10.0
google-auth==2.11.0
googleapis-common-protos==1.56.4
graphviz==0.20.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpagentparser==1.9.3
httpcore==0.15.0
httpx==0.23.0
idna==3.3
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==4.13.0
lazy-object-proxy==1.7.1
ldap3==2.9.1
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.0
mdurl==0.1.2
msal==1.18.0
msal-extensions==0.3.1
msrest==0.7.1
mypy-extensions==0.4.3
numpy==1.23.2
oauthlib==3.2.0
opencensus==0.11.0
opencensus-context==0.1.3
opencensus-ext-azure==1.1.7
packaging==21.3
pandas==1.4.4
paramiko==2.11.0
passlib==1.7.4
pathspec==0.9.0
pendulum==2.1.2
pgadmin4==6.13
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
portalocker==2.5.1
prison==0.2.1
protobuf==4.21.5
psutil==5.9.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
pymssql==2.2.5
PyNaCl==1.5.0
pyotp==2.6.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-daemon==2.3.1
python-dateutil==2.8.2
python-engineio==4.3.4
python-nvd3==0.15.0
python-slugify==6.1.2
python-socketio==5.7.1
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
qrcode==7.3.1
requests==2.28.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.5.1
rsa==4.9
s3transfer==0.5.2
setproctitle==1.3.2
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
speaklater3==1.4
SQLAlchemy==1.4.27
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.2
sshtunnel==0.4.0
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.3.0
ua-parser==0.16.1
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.11
user-agents==2.2.0
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==2.3.3
